chore: Update npm dependencies and add NestJS mapped types, class-transformer, and class-validator
This commit updates the npm dependencies in the `package.json` file. It adds the `@nestjs/config` and `@nestjs/typeorm` packages, and updates the `pg` and `typeorm` packages to their latest versions. Additionally, it adds the `@nestjs/mapped-types` package for NestJS mapped types, and the `class-transformer` and `class-validator` packages for data transformation and validation.